pyproject.toml
[project]
name = "langgrap-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=1.4.0",
    "langgraph-cli[inmem]>=0.4.19",
    "marker-pdf==1.5.5",
    "opencv-python-headless>=4.11.0.86,<5.0.0",
    "pillow>=10.3.0,<11.0.0",
    "python-dotenv>=1.2.2",
    "transformers>=4.45.2,<4.52.0",
    "torch>=2.5.1,<3.0.0",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg2-binary>=2.9.11",
]

[tool.uv]
override-dependencies = [
    "opencv-python; sys_platform == 'never'",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]

